refactor(architecture): import Problem/Result from canonical crate::observe in Layer 3
Layer 3 files imported `Problem`/`Result` from `crate::primitives::types`
(Layer 1) instead of the canonical Layer 3 path `crate::observe`, tightening
coupling to the internal primitives layer and bypassing the observe re-export
chain (primitives::types -> observe/problem/types.rs -> observe/mod.rs).

Rewrite the import path in all 25 affected Layer 3 files to
`use crate::observe::{Problem, Result}`, matching the established pattern in
security/paths/shortcuts.rs and siblings. The canonical re-export source
observe/problem/types.rs is deliberately left importing from primitives::types
(it is the single point that re-exports the Layer 1 definitions).

Closes #403
