+++ crates/octarine/src/auth/mfa/manager.rs
@@ -3,11 +3,11 @@
 //! Provides MFA operations with audit logging.
 
 use crate::observe;
+use crate::observe::Problem;
 use crate::primitives::auth::mfa::{
     RecoveryCodes, TotpCode, TotpConfig, TotpSecret, generate_recovery_codes, generate_totp_code,
     generate_totp_secret, get_otpauth_uri, validate_totp_code,
 };
-use crate::primitives::types::Problem;
 
 // ============================================================================
 // MFA Enrollment Result
